代码格式化 close #245 1.18.11[2.83] 第-10.4周 小寒/大寒2-2

+++ Assets/Scripts/Common/SqlCipher4Unity3D/Example/test/Editor/IgnoreTest.cs
@@ -14,46 +14,28 @@ namespace SqlCipher4Unity3D.Example.test.Editor
     {
         public string ID { get; set; }
         public string CompanyID { get; set; }
+
         [Ignore]
-        public Company Company
-        {
-            get
-            {
-                return new Company();
-            }
-            set
-            {
-                CompanyID = value.ID;
-            }
+        public Company Company {
+            get => new Company();
+            set => CompanyID = value.ID;
         }
 
         public override bool Equals(object obj)
         {
-            Account o = obj as Account;
-            if (o is null)
-            {
-                return false;
-            }
-
-            if (this.ID != o.ID)
-            {
-                return false;
-            }
-
-            if (this.CompanyID != o.CompanyID)
-            {
-                return false;
-            }
-
+            var o = obj as Account;
+            if (o is null) return false;
+            if (ID != o.ID) return false;
+            if (CompanyID != o.CompanyID) return false;
             return true;
         }
 
         public override int GetHashCode()
         {
             var hashCode = 480581749;
-            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(this.ID);
-            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(this.CompanyID);
-            hashCode = hashCode * -1521134295 + EqualityComparer<Company>.Default.GetHashCode(this.Company);
+            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(ID);
+            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(CompanyID);
+            hashCode = hashCode * -1521134295 + EqualityComparer<Company>.Default.GetHashCode(Company);
             return hashCode;
         }
     }
@@ -61,42 +43,35 @@ namespace SqlCipher4Unity3D.Example.test.Editor
     [TestFixture]
     public class IgnoreTest
     {
-        string DbPath = $"Assets/StreamingAssets/{nameof(IgnoreTest)}.db";
-        SQLiteConnection Conn = null;
+        private string DbPath = $"Assets/StreamingAssets/{nameof(IgnoreTest)}.db";
+        private SQLiteConnection Conn = null;
+
         [SetUp]
         public void Up()
         {
             Cleanup();
-            this.Conn = new SQLiteConnection(DbPath, "");
+            Conn = new SQLiteConnection(DbPath, "");
         }
 
         [TearDown]
         public void Down()
         {
-            if (Conn != null)
-            {
-                Conn.Close();
-            }
-
+            if (Conn != null) Conn.Close();
             Cleanup();
         }
 
-        void Cleanup()
+        private void Cleanup()
         {
-            if (File.Exists(DbPath))
-            {
-                File.Delete(DbPath);
-            }
+            if (File.Exists(DbPath)) File.Delete(DbPath);
         }
 
         [Test]
         public void IgnoreTestSimplePasses()
         {
             Conn.CreateTable<Account>();
-
-            Account a = new Account { ID = "A", CompanyID = "X" };
+            var a = new Account { ID = "A", CompanyID = "X" };
             Conn.Insert(a);
-            Account b = Conn.Table<Account>().Where(x => x.ID == "A").First();
+            var b = Conn.Table<Account>().Where(x => x.ID == "A").First();
             Assert.AreEqual(a, b);
         }
     }
